docs: keep README/index Quick start in sync; lead C with the tarball (#432)
Bring README.md's Python quick start up to date with docs/index.md's
(same install-first fix, no divergence between the two).

The C example's prose already said "grab a pre-built tarball" first,
but the shown compile command only worked after a full from-source
build (build/native/inc, build/libdoppler.a) -- exactly what the tip
was trying to help readers avoid. Point both the prose and the actual
command at a $PREFIX install layout instead (verified against a real
`cmake --install` output), with build-from-source called out as the
alternative rather than the default.

Also fixes a real, unrelated build warning caught while rebuilding:
examples/c/hbdecim_demo.c's printf passed N_TAPS (int) for a %zu format
specifier.

+++ examples/c/hbdecim_demo.c
@@ -50,7 +50,7 @@ main (void)
       return 1;
     }
 
-  printf ("Decimator created: %zu taps, rate=%.1f\n\n", N_TAPS,
+  printf ("Decimator created: %zu taps, rate=%.1f\n\n", (size_t)N_TAPS,
           HalfbandDecimator_get_rate (dec));
 
   /* Complex tone at f_n = 0.125 (eighth rate — well inside passband) */
